move virtual column and skip state to common
